+++ CHANGELOG.md
@@ -11,6 +11,7 @@ See `PUBLISH.md` for instructions on how to publish a new version.
 -->
 
 - (minor) Add styling for Markdown tables
+- (minor) Add syntax for columns to customise layout
 - (minor) Render captions for singleton images with titles